+++ src/uenv/mount.h
@@ -2,6 +2,7 @@
 
 #include <filesystem>
 #include <string>
+#include <sys/types.h>
 #include <vector>
 
 #include <fmt/core.h>
@@ -49,6 +50,15 @@ parse_and_validate_mounts(const std::string& description,
 /// called as root, in slurm-plugin
 util::expected<void, std::string> unshare_as_root();
 
+/// setup hook for the setuid squashfs-mount build: unshare the mount
+/// namespace and become the real root user, so that squashfs images can be
+/// mounted with the kernel driver.
+util::expected<void, std::string> unshare_and_become_root();
+
+/// post hook for the setuid squashfs-mount build: return to the calling
+/// user and disallow gaining any new privileges.
+util::expected<void, std::string> return_to_user_and_no_new_privs(uid_t uid);
+
 /// mount sqfs images, make sure mnt ns has been unshared before calling this
 /// function
 util::expected<void, std::string> do_mount(const mount_list& mount_entries);
@@ -67,6 +77,9 @@ mount_tmpfs(std::filesystem::path dst, std::optional<std::uint64_t> size);
 /// bind mount
 util::expected<void, std::string> bind_mount(std::filesystem::path src,
                                              std::filesystem::path dst);
+
+// make mutable root (inspired by bwrap)
+util::expected<void, std::string> make_mutable_root();
 
 } // namespace uenv
