+++ src/uenv/rootless.cpp
@@ -9,153 +9,18 @@ extern "C" {
 }
 #include "macros.h"
 #include "posix_io.h"
-#include "rootless.h"
-#include <semaphore.h>
 #include <spdlog/spdlog.h>
 #include <stddef.h>
 #include <string>
-#include <sys/mman.h>
 #include <sys/mount.h>
-#include <sys/prctl.h>
 #include <uenv/mount.h>
 #include <unistd.h>
 #include <util/expected.h>
-
-// timeout in seconds for waiting for join semaphore.
-#define JOIN_TIMEOUT 30
+#include <fcntl.h>
+#include <sys/prctl.h>
 
 namespace uenv {
-
-// Join a specific namespace.
-util::expected<void, std::string> namespace_join(pid_t pid, const char* ns) {
-    std::string path;
-    int fd;
-
-    path = fmt::format("/proc/{}/ns/{}", pid, ns);
-    fd = open(path.c_str(), O_RDONLY);
-    if (fd == -1) {
-        if (errno == ENOENT)
-            return util::unexpected(
-                fmt::format("join: no PID {}: {} not found", pid, path));
-        else
-            return util::unexpected(fmt::format("join: can't open {}", path));
-    }
-    // setns(2) seems to be involved in some kind of race with syslog(3).
-    // Rarely, when configured with --enable-syslog, the call fails with
-    // EINVAL. We never figured out a proper fix, so just retry a few times in
-    // a loop. See issue #1270.
-    for (int i = 1; setns(fd, 0) != 0; i++)
-        if (i >= 5) {
-            return util::unexpected(
-                fmt::format("can’t join {} namespace of pid {}", ns, pid));
-        } else {
-            spdlog::warn("can’t join {} namespace; trying again", ns);
-            sleep(1);
-        }
-    return {};
-}
-
-// Join the existing namespaces containing process pid, which could be the
-// join winner or another process.
-util::expected<void, std::string> namespaces_join(pid_t pid) {
-    spdlog::warn("joining namespaces of pid {}", pid);
-    if (auto r = namespace_join(pid, "user"); !r) {
-        return r;
-    }
-    if (auto r = namespace_join(pid, "mnt"); !r) {
-        return r;
-    }
-    return {};
-}
-
-// Wait for semaphore sem for up to timeout seconds. If timeout or an error,
-// exit unsuccessfully.
-util::expected<void, std::string> sem_timedwait_relative(sem_t* sem,
-                                                         int timeout) {
-    struct timespec deadline;
-
-    // sem_timedwait() requires a deadline rather than a timeout.
-    Z_e(clock_gettime(CLOCK_REALTIME, &deadline));
-    deadline.tv_sec += timeout;
-    Zfe(sem_timedwait(sem, &deadline), "failure waiting for join lock");
-    return {};
-}
-
-// Begin coordinated section of namespace joining.
-util::expected<void, std::string> join_begin(join_t& join,
-                                             std::string join_tag) {
-    int fd;
-    join.sem_name = fmt::format("/uenv-run_sem-{}", join_tag);
-    join.shm_name = fmt::format("/uenv-run_shm-{}", join_tag);
-
-    // Serialize.
-    join.sem = sem_open(join.sem_name.c_str(), O_CREAT, 0600, 1);
-    T_e(join.sem != SEM_FAILED);
-    if (auto r = sem_timedwait_relative(join.sem, JOIN_TIMEOUT); !r)
-        return r;
-
-    // Am I the winner?
-    fd = shm_open(join.shm_name.c_str(), O_CREAT | O_EXCL | O_RDWR, 0600);
-    if (fd > 0) {
-        spdlog::trace("join:: I won! PID {}", getpid());
-        join.winner_p = true;
-        Z_e(ftruncate(fd, sizeof(*join.shared)));
-    } else {
-        std::string err{strerror(errno)};
-        T_e(errno == EEXIST);
-        spdlog::trace("join: I lost {}", err);
-        join.winner_p = false;
-        fd = shm_open(join.shm_name.c_str(), O_RDWR, 0);
-        T_e(fd > 0);
-    }
-
-    join.shared = static_cast<decltype(join.shared)>(mmap(
-        NULL, sizeof(*join.shared), PROT_READ | PROT_WRITE, MAP_SHARED, fd, 0));
-    T__(join.shared != NULL);
-    Z_e(close(fd));
-
-    // Winner keeps lock; losers parallelize (winner will be done by now).
-    if (!join.winner_p)
-        Z_e(sem_post(join.sem));
-    return {};
-}
-
-// End coordinated section of namespace joining.
-util::expected<void, std::string> join_end(join_t& join, int join_ct,
-                                           std::optional<pid_t> winner_pid) {
-    if (join.winner_p) { // winner still serial
-        spdlog::trace("join: winner initializing shared data");
-        if (!winner_pid)
-            join.shared->winner_pid = getpid();
-        else
-            join.shared->winner_pid = winner_pid.value();
-        join.shared->proc_left_ct = join_ct;
-    } else // losers serialize
-        sem_timedwait_relative(join.sem, JOIN_TIMEOUT);
-
-    join.shared->proc_left_ct--;
-    spdlog::trace("join: {} peers left excluding myself",
-                  join.shared->proc_left_ct);
-
-    if (join.shared->proc_left_ct <= 0) {
-        spdlog::trace("join: cleaning up IPC resources");
-        Tf_(join.shared->proc_left_ct == 0,
-            "join: expected 0 peers left but found {}",
-            join.shared->proc_left_ct);
-        Zfe(sem_unlink(join.sem_name.c_str()), "join: can't unlink sem: {}",
-            join.sem_name.c_str());
-        Zfe(shm_unlink(join.shm_name.c_str()), "join: can't unlink shm: {}",
-            join.shm_name.c_str());
-    }
-
-    Z_e(sem_post(join.sem)); // parallelize (all)
-
-    Z_e(munmap(join.shared, sizeof(*join.shared)));
-    Z_e(sem_close(join.sem));
-
-    spdlog::trace("join: done");
-    return {};
-}
+namespace rootless {
 
 // Same effect as `unshare --mount --map-root-user`
 util::expected<void, std::string> unshare_mount_map_root() {
@@ -182,7 +47,7 @@ util::expected<void, std::string> unshare_mount_map_root() {
     if (auto r = write(proc_uid_map.value(), buf, strlen(buf)); !r) {
         return r;
     }
-    T_e(close(proc_uid_map.value()));
+    Z_e(close(proc_uid_map.value()));
 
     // write /proc/self/gid_setgroups -> deny
     auto proc_setgroups = openat(AT_FDCWD, "/proc/self/setgroups", O_WRONLY);
@@ -191,7 +56,7 @@ util::expected<void, std::string> unshare_mount_map_root() {
     if (auto r = write(proc_setgroups.value(), "deny", 4); !r) {
         return r;
     }
-    T_e(close(proc_setgroups.value()));
+    Z_e(close(proc_setgroups.value()));
 
     // map gid  to root group
     auto proc_gid_map = openat(AT_FDCWD, "/proc/self/gid_map", O_WRONLY);
@@ -201,7 +66,7 @@ util::expected<void, std::string> unshare_mount_map_root() {
     if (auto r = write(proc_gid_map.value(), buf, strlen(buf)); !r) {
         return r;
     }
-    T_e(close(proc_gid_map.value()));
+    Z_e(close(proc_gid_map.value()));
 
     // the following is executed by `unshare --mount --map-root-user`
     if (auto r = mount("none", "/", std::nullopt, MS_REC | MS_PRIVATE, nullptr);
@@ -217,6 +82,7 @@ util::expected<void, std::string> map_effective_user(uid_t uid, gid_t gid) {
     Z_e(unshare(CLONE_NEWUSER | CLONE_NEWNS));
     // map current user id to root
     char buf[256];
+    spdlog::trace("map_effective_user({}, {})", uid, gid);
     auto proc_uid_map = openat(AT_FDCWD, "/proc/self/uid_map", O_WRONLY);
     if (!proc_uid_map)
         return util::unexpected(proc_uid_map.error());
@@ -224,15 +90,12 @@ util::expected<void, std::string> map_effective_user(uid_t uid, gid_t gid) {
     if (auto r = write(proc_uid_map.value(), buf, strlen(buf)); !r) {
         return r;
     }
-    T_e(close(proc_uid_map.value()));
+    Z_e(close(proc_uid_map.value()));
 
-    auto proc_setgroups = openat(AT_FDCWD, "/proc/self/setgroups", O_WRONLY);
-    if (!proc_setgroups)
-        return util::unexpected(proc_setgroups.error());
-    if (auto r = write(proc_setgroups.value(), "allow", 5); !r) {
-        return r;
-    }
-    T_e(close(proc_setgroups.value()));
+    // note: setgroups is already "deny" here, inherited from the enclosing
+    // fake-root namespace (unshare_mount_map_root) -- once a namespace's
+    // setgroups is "deny", all descendant namespaces are permanently locked
+    // to "deny" too, so writing "allow" here would always fail with EPERM.
 
     auto proc_gid_map = openat(AT_FDCWD, "/proc/self/gid_map", O_WRONLY);
     if (!proc_gid_map)
@@ -241,10 +104,21 @@ util::expected<void, std::string> map_effective_user(uid_t uid, gid_t gid) {
     if (auto r = write(proc_gid_map.value(), buf, strlen(buf)); !r) {
         return r;
     }
-    T_e(close(proc_gid_map.value()));
+    Z_e(close(proc_gid_map.value()));
 
     // disable coredump again (slurm policy)
     Z_e(prctl(PR_SET_DUMPABLE, 0));
+    return {};
+}
+
+util::expected<void, std::string> exit_sandbox(uid_t uid, gid_t gid) {
+    if (auto r = map_effective_user(uid, gid); !r) {
+        return r;
+    }
+
+    if (prctl(PR_SET_NO_NEW_PRIVS, 1, 0, 0, 0) != 0) {
+        return util::unexpected("PR_SET_NO_NEW_PRIVS failed");
+    }
     return {};
 }
 
@@ -284,8 +158,7 @@ static void init_fs_ops(struct fuse_lowlevel_ops* sqfs_ll_ops) {
 // Ref: https://github.com/vasi/squashfuse/blob/master/ll_main.c
 util::expected<void, std::string> do_sqfs_ll_mount(const mount_pair& entry,
                                                    bool fuse_st) {
-
-    spdlog::trace("do_sqfs_mount");
+    spdlog::trace("do_sqfs_ll_mount");
     // use a pipe to synchronize parent and child process
     int pipe_wait[2];
     Z_e(pipe(pipe_wait));
@@ -437,5 +310,5 @@ util::expected<void, std::string> do_sqfs_ll_mount(const mount_pair& entry,
 
     return {};
 }
-
+}  // rootless
 } // namespace uenv
