+++ src/uenv/rootless.h
@@ -1,21 +1,9 @@
 #pragma once
 
 #include "uenv/mount.h"
-#include <semaphore.h>
 
 namespace uenv {
-
-/* Variables for coordinating join */
-struct join_t {
-    bool winner_p;
-    std::string sem_name;
-    sem_t* sem;
-    std::string shm_name;
-    struct {
-        pid_t winner_pid;
-        int proc_left_ct; // serial-only access
-    }* shared;
-};
+namespace rootless {
 
 // Same effect as `unshare --mount --map-root-user`
 util::expected<void, std::string> unshare_mount_map_root();
@@ -23,17 +11,13 @@ util::expected<void, std::string> unshare_mount_map_root();
 // go back to effective user
 util::expected<void, std::string> map_effective_user(uid_t uid, gid_t gid);
 
+// post hook for the rootless squashfs-mount build: exit the fake-root user
+// namespace and disallow gaining any new privileges.
+util::expected<void, std::string> exit_sandbox(uid_t uid, gid_t gid);
+
 // mount using squashfuse low-level interface
 util::expected<void, std::string> do_sqfs_ll_mount(const uenv::mount_pair&,
                                                    bool fuse_st);
 
-util::expected<void, std::string> make_mutable_root();
-
-util::expected<void, std::string> join_begin(join_t& join,
-                                             std::string join_tag);
-util::expected<void, std::string> join_end(join_t& join, int join_ct,
-                                           std::optional<pid_t> winner_pid);
-
-util::expected<void, std::string> namespaces_join(pid_t pid);
-
+} // namespace rootless
 } // namespace uenv
